move middlewares to titiler.core
